+++ server/app/assets/javascripts/admin_entry_point.ts
@@ -18,7 +18,7 @@ import * as adminSettingsView from './admin_settings_view'
 import * as adminValidation from './admin_validation'
 import * as apiDocs from './api_docs'
 import * as devIcons from './dev_icons'
-import * as map from './map'
+import * as map from './mapquestion/map'
 import * as modal from './modal'
 import * as questionBank from './questionBank'
 import PreviewController, * as preview from './preview'
